Modularize Logger
