--- lib/logger_ext.rb
@@ -1,11 +1,0 @@
-# coding: utf-8
-
-if RUBY_VERSION >= '2.0.0'
-  module LogDeviceStdErr
-    def write(message)
-      STDERR.puts(message)
-      super(message)
-    end
-  end
-  Logger::LogDevice.send(:prepend, LogDeviceStdErr)
-end
